restructuring a lot
