+++ controllers/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  UsersController: require('./users.controller')
+}
