+++ controllers/users.controller.js
@@ -6,31 +6,12 @@ class UsersController {
   static index (req, res, next) {
     User.index()
       .then(users => {
+        console.log(users)
         return res.json({ users })
       })
       .catch(err => {
-        console.log('Error!', err);
+        console.log('Error!', err)
       })
-  }
-
-  static getById (req, res, next) {
-
-  }
-
-  static login (req, res, next) {
-
-  }
-
-  static create (req, res, next) {
-
-  }
-
-  static update (req, res, next) {
-
-  }
-
-  static destroy (req, res, next) {
-
   }
 
 }
