+++ db/knex.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const environment = process.env.NODE_ENV || 'development';
+const knexConfig = require('../knexfile')[environment];
+const knex = require('knex')(knexConfig);
+
+module.exports = knex;
