+++ knexfile.js
@@ -1,0 +1,39 @@
+'use strict'
+const path = require('path')
+
+module.exports = {
+
+  development: {
+    client: 'pg',
+    connection: 'postgres://localhost:5432/commit_m',
+    migrations: {
+      directory: path.join(__dirname, 'db', 'migrations')
+    },
+    seeds: {
+      directory: path.join(__dirname, 'db', 'seeds')
+    }
+  },
+
+
+  test: {
+    client: 'pg',
+    connection: 'postgres://localhost/commit_m',
+    migrations: {
+      directory: path.join(__dirname, 'db', 'migrations')
+    },
+    seeds: {
+      directory: path.join(__dirname, 'db', 'seeds')
+    }
+  },
+
+  production: {
+    client: 'pg',
+    connection: process.env.DATABASE_URL,
+    migrations: {
+      directory: path.join(__dirname, 'db', 'migrations')
+    },
+    seeds: {
+      directory: path.join(__dirname, 'db', 'seeds')
+    }
+  }
+}
