+++ package.json
@@ -15,10 +15,12 @@
     "bcrypt": "^1.0.3",
     "body-parser": "^1.18.2",
     "cors": "^2.8.4",
+    "dotenv": "^5.0.1",
     "express": "^4.16.3",
     "jsonwebtoken": "^8.2.0",
     "knex": "^0.14.4",
     "morgan": "^1.9.0",
+    "path": "^0.12.7",
     "uuid": "^3.2.1"
   },
   "devDependencies": {
