+++ routes/users.routes.js
@@ -1,0 +1,9 @@
+const express = require('express')
+const router = express.Router()
+const { UsersController } = require('../controllers')
+
+
+router.get('/all', UsersController.index)
+
+
+module.exports = router
