--- src/controllers/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  UsersController: require('./users.controller'),
-}
