--- src/models/users.model.js
@@ -1,38 +1,0 @@
-// const db = require('../db/knex.js')
-// const knex = require('../db/knex')
-// const bcrypt = require('bcryptjs')
-
-class User {
-  constructor() {}
-
-  static index() {
-    return true
-  }
-
-  static getById() {
-
-  }
-
-  static getUserByEmail () {
-
-  }
-
-  static tryLoginUser() {
-
-  }
-
-  static create() {
-
-  }
-
-  static update() {
-
-  }
-
-  static destroy() {
-
-  }
-
-}
-
-module.exports = User
