--- src/routes/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  UsersRouter: require('./users.routes'),
-}
